+++ entity.h
@@ -19,7 +19,7 @@ public:
 	* @param v velocity
 	* @param t texture
 	*/
-	Entity(Vector s, Vector p, int l, Vector v, SDL_Texture* t);
+	Entity(Vector s, Vector p, float l, Vector v, SDL_Texture* t);
 
 	/**
 	* Gets the entities xy-size values.
@@ -47,12 +47,12 @@ public:
 	* Gets the entities layer value.
 	* @return layer int of entity
 	*/
-	int getLayer(void);
+	float getLayer(void);
 	/**
 	* Sets the entities layer value to the given int.
 	* @param l layer int to set
 	*/
-	void setLayer(int l);
+	void setLayer(float l);
 
 	/**
 	* Gets the entities xy-velocity values to the given vector.
@@ -87,7 +87,7 @@ private:
 	/** Vector position of entity */
 	Vector position;
 	/** RenderLayer of entity. (higher layer = render on top) */
-	int layer = 0;
+	float layer = 0;
 	/** Vector velocity of entity */
 	Vector velocity;
 	/** Texture of entity */
